#88 getting the modules functional as route + controller separations.
